+++ packages/forma-36-react-components/src/components/Card/InlineEntryCard/InlineEntryCard.stories.tsx
@@ -1,111 +1,68 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { text, select, boolean } from '@storybook/addon-knobs';
-import { action } from '@storybook/addon-actions';
-
-import InlineEntryCard from './InlineEntryCard';
+import InlineEntryCard, { InlineEntryCardPropTypes } from './InlineEntryCard';
 import DropdownListItem from '../../Dropdown/DropdownListItem';
 import DropdownList from '../../Dropdown/DropdownList';
 
-storiesOf('Components/Card/InlineEntryCard', module)
-  .addParameters({
-    propTypes: InlineEntryCard['__docgenInfo'],
-    component: InlineEntryCard,
-  })
-  .add('default', () => (
-    <div
-      style={{
-        lineHeight: 1.5,
-        display: 'inline-block',
-        maxWidth: 600,
-        fontSize: 16,
-      }}
-    >
-      Fusce a odio pharetra, porta justo vel, maximus ex. In pellentesque a orci
-      et pretium. Praesent libero lorem, gravida eu pulvinar id, eleifend a
-      sapien. &nbsp;
-      <InlineEntryCard
-        className={text('className', '')}
-        isLoading={boolean('isLoading', false)}
-        isSelected={boolean('isSelected', false)}
-        status={select(
-          'status',
-          {
-            published: 'published',
-            draft: 'draft',
-            archived: 'archived',
-            changed: 'changed',
-          },
-          'published',
-        )}
-        dropdownListElements={
-          <DropdownList>
-            <DropdownListItem onClick={action('Edit onClick')}>
-              Edit
-            </DropdownListItem>
-            <DropdownListItem onClick={action('Download onClick')}>
-              Download
-            </DropdownListItem>
-            <DropdownListItem onClick={action('Remove onClick')}>
-              Remove
-            </DropdownListItem>
-          </DropdownList>
-        }
-      >
-        {text('children', 'Title of inline entry')}
-      </InlineEntryCard>
-      &nbsp; Nulla a ultrices nulla, vel blandit sapien. Etiam eget massa dui.
-      Fusce id nisl quam. Integer nec mi arcu. Nullam lacinia est lectus, a
-      euismod purus eleifend id. Fusce a odio pharetra, porta justo vel, maximus
-      ex. In pellentesque a orci et pretium. Praesent libero lorem, gravida eu
-      pulvinar id, eleifend a sapien. &nbsp;
-      <InlineEntryCard
-        className={text('className', '')}
-        isLoading={boolean('isLoading', false)}
-        isSelected={boolean('isSelected', false)}
-        status={select(
-          'status',
-          {
-            published: 'published',
-            draft: 'draft',
-            archived: 'archived',
-            changed: 'changed',
-          },
-          'published',
-        )}
-        dropdownListElements={
-          <React.Fragment>
-            <DropdownList>
-              <DropdownListItem isTitle>Actions</DropdownListItem>
-              <DropdownListItem onClick={action('Edit onClick')}>
-                Edit
-              </DropdownListItem>
-              <DropdownListItem onClick={action('Download onClick')}>
-                Download
-              </DropdownListItem>
-              <DropdownListItem onClick={action('Remove onClick')}>
-                Remove
-              </DropdownListItem>
-            </DropdownList>
-            <DropdownList>
-              <DropdownListItem isTitle>Actions</DropdownListItem>
-              <DropdownListItem onClick={action('Edit onClick')}>
-                Edit
-              </DropdownListItem>
-              <DropdownListItem onClick={action('Download onClick')}>
-                Download
-              </DropdownListItem>
-              <DropdownListItem onClick={action('Remove onClick')}>
-                Remove
-              </DropdownListItem>
-            </DropdownList>
-          </React.Fragment>
-        }
-      >
-        {text('children', 'Title of inline entry')}
-      </InlineEntryCard>
-      &nbsp; Nulla a ultrices nulla, vel blandit sapien. Etiam eget massa dui.
-      Fusce id nisl quam. Integer nec mi arcu. Nullam lacinia est lectus, a
-      euismod purus eleifend id.
-    </div>
-  ));
+export default {
+  title: 'Components/Card/InlineEntryCard',
+  component: InlineEntryCard,
+  parameters: {
+    propTypes: [InlineEntryCard['__docgenInfo']],
+  },
+  argTypes: {
+    className: { control: { disable: true } },
+    testId: { control: { disable: true } },
+  },
+};
+
+const dropdownElements = (
+  <React.Fragment>
+    <DropdownList>
+      <DropdownListItem isTitle>Actions</DropdownListItem>
+      <DropdownListItem href="#">Edit (with href)</DropdownListItem>
+      <DropdownListItem onClick={() => {}}>Download</DropdownListItem>
+      <DropdownListItem onClick={() => {}}>Remove</DropdownListItem>
+    </DropdownList>
+    <DropdownList>
+      <DropdownListItem isTitle>Actions</DropdownListItem>
+      <DropdownListItem onClick={() => {}}>Edit</DropdownListItem>
+      <DropdownListItem onClick={() => {}}>Download</DropdownListItem>
+      <DropdownListItem onClick={() => {}}>Remove</DropdownListItem>
+    </DropdownList>
+  </React.Fragment>
+);
+
+interface Args extends InlineEntryCardPropTypes {
+  entryText?: string;
+}
+
+export const Basic = ({ entryText, ...args }: Args) => (
+  <div
+    style={{
+      lineHeight: 1.5,
+      display: 'inline-block',
+      maxWidth: 600,
+      fontSize: 16,
+    }}
+  >
+    Fusce a odio pharetra, porta justo vel, maximus ex. In pellentesque a orci
+    et pretium. Praesent libero lorem, gravida eu pulvinar id, eleifend a
+    sapien. &nbsp;
+    <InlineEntryCard {...args}>{entryText}</InlineEntryCard>
+    &nbsp; Nulla a ultrices nulla, vel blandit sapien. Etiam eget massa dui.
+    Fusce id nisl quam. Integer nec mi arcu. Nullam lacinia est lectus, a
+    euismod purus eleifend id. Fusce a odio pharetra, porta justo vel, maximus
+    ex. In pellentesque a orci et pretium. Praesent libero lorem, gravida eu
+    pulvinar id, eleifend a sapien. &nbsp;
+    <InlineEntryCard {...args}>{entryText}</InlineEntryCard>
+    &nbsp; Nulla a ultrices nulla, vel blandit sapien. Etiam eget massa dui.
+    Fusce id nisl quam. Integer nec mi arcu. Nullam lacinia est lectus, a
+    euismod purus eleifend id.
+  </div>
+);
+
+Basic.args = {
+  status: 'published',
+  dropdownListElements: dropdownElements,
+  entryText: 'Title of inline entry',
+};