+++ packages/forma-36-react-components/src/components/Illustration/Illustration.stories.tsx
@@ -1,46 +1,51 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { text, select } from '@storybook/addon-knobs';
 
-import Illustration, { IllustrationType } from './Illustration';
+import Illustration, {
+  IllustrationProps,
+  IllustrationType,
+} from './Illustration';
 import { illustrationName } from './constants';
+import Flex from '../Flex/Flex';
+import Grid from '../Grid/Grid';
 
-storiesOf('Components/Illustration', module)
-  .addParameters({
-    propTypes: Illustration['__docgenInfo'],
-    component: Illustration,
-  })
-  .add('default', () => (
-    <Illustration
-      illustration={select(
-        'illustration',
-        Object.keys(illustrationName),
-        Object.keys(illustrationName)[0],
-      )}
-      className={text('className', '')}
-    />
-  ))
-  .add('All Illustrations', () => (
-    <div
-      style={{
-        columns: '3',
-        width: '800px',
-      }}
-    >
-      {Object.keys(illustrationName)
-        .filter(
-          (illustration) => !illustration.toLowerCase().includes('trimmed'),
-        )
-        .map((illustration) => (
-          <div
-            key={illustration}
-            style={{ padding: '4px', fontSize: '14px', lineHeight: '21px' }}
-          >
+export default {
+  title: 'Components/Illustration',
+  component: Illustration,
+  parameters: {
+    propTypes: [Illustration['__docgenInfo']],
+  },
+  argTypes: {
+    className: { control: { disable: true } },
+    testId: { control: { disable: true } },
+  },
+};
+
+export const Basic = (args: IllustrationProps) => <Illustration {...args} />;
+
+Basic.args = {
+  illustration: 'Archive',
+};
+
+export const AllIllustrations = () => (
+  <Grid columns={'auto 1fr 1fr'} columnGap="spacingM">
+    {Object.keys(illustrationName)
+      .filter((illustration) => !illustration.toLowerCase().includes('trimmed'))
+      .map((illustration) => (
+        <Flex
+          key={illustration}
+          padding="spacingS"
+          marginRight="spacingM"
+          alignItems="center"
+          justifyContent="flex-start"
+          flexGrow={0}
+        >
+          <Flex marginRight="spacingS">
             <Illustration illustration={illustration as IllustrationType} />
-            <span style={{ display: 'inline-block', verticalAlign: 'middle' }}>
-              {illustration}
-            </span>
-          </div>
-        ))}
-    </div>
-  ));
+          </Flex>
+          <span style={{ display: 'inline-block', verticalAlign: 'middle' }}>
+            {illustration}
+          </span>
+        </Flex>
+      ))}
+  </Grid>
+);