+++ packages/forma-36-react-components/src/components/Typography/DisplayText/DisplayText.tsx
@@ -6,7 +6,7 @@ import { TypographyContext } from '../Typography';
 
 export interface DisplayTextProps {
   element?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';
-  size?: 'default' | 'large';
+  size?: 'default' | 'large' | 'huge';
   className?: string;
   children?: React.ReactNode;
   style?: React.CSSProperties;