+++ packages/forma-36-react-components/src/components/Typography/Paragraph/Paragraph.stories.tsx
@@ -1,23 +1,25 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { text, select } from '@storybook/addon-knobs';
+import Paragraph, { ParagraphProps } from './Paragraph';
 
-import Paragraph from './Paragraph';
+export default {
+  title: 'Typography/Paragraph',
+  component: Paragraph,
+  parameters: {
+    propTypes: [Paragraph['__docgenInfo']],
+  },
+  argTypes: {
+    className: { control: { disable: true } },
+  },
+};
 
-storiesOf('Components/Typography/Paragraph', module)
-  .addParameters({
-    propTypes: Paragraph['__docgenInfo'],
-    component: Paragraph,
-  })
-  .add('default', () => (
-    <Paragraph
-      className={text('className', '')}
-      element={select(
-        'element',
-        ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'p'],
-        'p',
-      )}
-    >
-      {text('children', 'Paragraph')}
-    </Paragraph>
-  ));
+interface Args extends ParagraphProps {
+  paragraphText: string;
+}
+
+export const Basic = ({ paragraphText, ...args }: Args) => (
+  <Paragraph {...args}>{paragraphText}</Paragraph>
+);
+
+Basic.args = {
+  paragraphText: 'Paragraph',
+};